+++ Tests/StencilTests/EnvironmentSpec.swift
@@ -77,14 +77,14 @@ final class EnvironmentTests: XCTestCase {
     }
 
     it("reports syntax error on invalid filter expression") {
-      self.template = "{{ array|filterEach: $0|isPositive }}"
+      self.template = "{{ array|select: $0|isPositive }}"
       let ext = Extension()
       ext.registerFilter("isPositive") { (value) -> Any? in
         if let number = toNumber(value: value as Any) { return number > 0 }
         else { return nil }
       }
       self.environment = Environment(extensions: [ext])
-      try self.expectError(reason: "Can't parse filter expression", token: "array|filterEach: $0|isPositive")
+      try self.expectError(reason: "Can't parse filter expression", token: "array|select: $0|isPositive")
     }
   }