+++ Tests/StencilTests/FilterSpec.swift
@@ -426,7 +426,7 @@ final class FilterTests: XCTestCase {
     }
   }
 
-  func testFilterEachFilter() throws {
+  func testSelectFilter() throws {
     it("can filter using filter") {
       let ext = Extension()
       ext.registerFilter("isPositive") { (value) -> Any? in
@@ -435,19 +435,19 @@ final class FilterTests: XCTestCase {
       }
       let env = Environment(extensions: [ext])
 
-      let template = Template(templateString: #"{{ array|filterEach:"$0|isPositive" }}"#)
+      let template = Template(templateString: #"{{ array|select:"$0|isPositive" }}"#)
       let result = try template.render(Context(dictionary: ["array": [1, -1, 2, -2, 3, -3]], environment: env))
       try expect(result) == "[1, 2, 3]"
     }
 
     it("can filter using boolean expression") {
-      let template = Template(templateString: #"{{ array|filterEach:"$0 > 0" }}"#)
+      let template = Template(templateString: #"{{ array|select:"$0 > 0" }}"#)
       let result = try template.render(Context(dictionary: ["array": [1, -1, 2, -2, 3, -3]]))
       try expect(result) == "[1, 2, 3]"
     }
 
     it("can filter using variable expression") {
-      let template = Template(templateString: "{{ array|filterEach:\"$0\" }}")
+      let template = Template(templateString: "{{ array|select:\"$0\" }}")
       let result = try template.render(Context(dictionary: ["array": [1, -1, nil, 2, -2, 3, -3]]))
       try expect(result) == "[1, 2, 3]"
     }